build(npm): whitelist files instead of using npmignore
SECURITY DISCLOSURE: As part of v0.30.0, I inadvertently published a file `.envrc` containing a personal GitHub token with write access to this repository used for release automation to npm. I immediately revoked the token on discovering its publication, at which point it had been public for ~2h. After a manual review (given the limited scope of the token), I am confident the leaked token was not been used maliciously in the window of time it was public.

To prevent similar mistakes in future, this commit removes .npmignore in favor of whitelisting the built TypeScript files in /dist via the "files" property in package.json.